Revert "orchagent: Updating the route next hop ID also sets action to forward" (#138)
This change breaks the latest test as the current SAI implementation doesn't support
the mid-state that the route with packet action DROP is set with next hop ID and packet
action FORWARD by two steps. The implementation also prevents removing the default route
and create a new route instead. After the SAI implementation is fixed, this change
will be re-introduced.

ref: https://github.com/opencomputeproject/SAI/pull/253