[orchagent]: Support port/LAG in multiple VLANs (#327)
One port/LAG may join multi-VLAN in tagged/untagged mode.
pvid on LAG is not supported in SAI v1.0, as workaround we deal with
each LAG member explicitly for pvid. Bridge port has to be created
first for port/LAG to join 1Q VLAN. The same bridge port is used to join all VLANs.
For each VLAN that port/LAG has joined, there is one vlan member id
which is the handle for VLAN member attribute get/set and remove.