[portmgrd]: Add portmgrd to monitor port MTU configurations (#545)
In order to get rid of /etc/network/interfaces file and move
all the configurations to the configuration database, MTU configurations
are required to be put into the database and monitored by a specific
daemon. This daemon portmgrd is to achieve this goal.

Currently, this daemon will only listen to the port MTU configurations
if existed in the database and then call the 'ip link' command to
configure the kernel netdevs.

It will also capture the admin status set in the configuration database
and call the 'ip link' command to configure the kernel netdevs.

The default MTUs, however, are set as the default value in the orchagent
port.h file. The MTU value in the netlink message is anyway ignored.

In the next stage, the change of the MTU will be supported so that the
kernel host interface MTU will be reflected to the SAI port/router
interface MTU.

Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>