[dvs] Update test_nhg.py to make it pass in synchronous mode (#1438)
Add a wait for route entries in ASIC_DB in test_nhg.py::TestNextHopGroup::test_route_nhg_exhaust so that the test case could pass in synchronous mode.

The test case test_nhg.py::TestNextHopGroup::test_route_nhg_exhaust fails in synchronous mode due to a delay between the creation of next-hop groups and routes in ASIC_DB. Getting the count of routes entries immediately after the creation of next-hop groups results in an inaccurate result. This PR fixes this by waiting for the number of route entries in ASIC_DB to be the expected value.