[pfcwd] Update PFC storm detection logic for Mellanox platforms (#1586)
Use "PFC duration" counters in micro seconds instead of quanta

SONiC PFCWD logic requires "pfc duration" value in micro seconds but in
SAI it was provided as quanta of time. So it required additional conversion
which used speed value to do such conversion and it could cause PFCWD to
detect storm on operationally down port in case of link flapping.

Now there are new SAI attributes that provide "pfc duration" in micro
seconds so PCWD storm detection logic is updated in order to use this
new "pfc duration" counters. Such algorithm change helps to avoid false
PFC storm detection in case of link flapping because conversion is not
needed anymore.

Signed-off-by: Volodymyr Samotiy <volodymyrs@nvidia.com>