Stablize the test case (#1679)
- What I did
Stabilize the vs test.

- Why I did it
Stabilize the vs test.

- How I verified it
Run the vs test.

- Details if related
One logic in the vs test is to check consistency between APPL_DB and ASIC_DB for buffer profiles. However, the mapping between are stored in the orchagent and can't be accessed from outside. The way we fetch that mapping is:

Get the SAI OID of all the buffer profiles in ASIC_DB at the beginning of the test, and store it to set P1
Get the SAI OID of all the buffer profiles in ASIC_DB after a new buffer profile has been created, and stored it to P2
The newly created buffer profile in ASIC_DB should be P2 - P1.
However, sometimes there can be more than one OIDs in P2 - P1. This is because the old profile hasn't been removed from the ASIC, which typically caused by timing issues, which fails the test.
To make the test case stable, we will retry for 5 seconds to make sure the old profile is removed and the OID of the new profile can be fetched.

Signed-off-by: Stephen Sun <stephens@nvidia.com>