Update pool sizes during initialization from timer only (#1708)
- What I did
This is a supplement of #1685. In that PR, most of the cases in which the buffer pool updated during initialization have been avoided but there are still some rare cases not covered. Eg., it will try to recalculate the pool size because the lossless PSs are removed if one port is admin down.

- Why I did it
Make sure that the the buffer pools will be updated only once during initialization.

- How I verified it
Run regression and manual test

Signed-off-by: Stephen Sun <stephens@nvidia.com>