[Bulk mode] Limit the size of bulker (#1744)
What I did
Limit the size of the bulker.

Why I did it
Without a constraint on the size of bulkers, the bulked operation may take longer than the SAI timeout to finish. Such behavior would result in errors and swss crashes. This PR limits the bulk size and avoids SAI timeout in bulk mode.