[Bulk mode] Limit the size of bulker (#1760)
What I did
This is to backport changes in #1744 to 202012.
Limit the size of the bulker.

Why I did it
Without a constraint on the size of bulkers, the bulked operation may take longer than the SAI timeout to finish. Such behavior would result in errors and swss crashes. This PR limits the bulk size and avoids SAI timeout in bulk mode.