[qosorch] Dot1p map list initialization fix (#1746)
- For DOT1P QOS map, the map list was not initialized correctly. So
"color" field in the key was sending garbage. Because of this, orchagent
serialization function is unable to translate and serialize the attribute
value properly and returns warning. During subsequent operation which comes
with a different garbage, this was treated as a different "color" and
syncd/SAI returns error saying "modification not implemented". This is
fixed by initializing the dotp1 map list attribute.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>