[request parser] Allow request parser to parse multiple values
What I did
Add an option for request parser to parse strings that contains multiple IP address, MAC address, or unsigned int. Return the values in a vector.

Why I did it
To enable overlay ECMP, the request may contain values with multiple nexthops. This PR aims to properly parse the request in such a scenario.