SAI_REDIS_SWITCH_ATTR_CONTEXT shouldbe the last attribute. This is what sairedis library expects (#1935)
**What I did**
Moved the code so that SAI_REDIS_SWITCH_ATTR_CONTEXT  is the last attribute in create switch call.

**Why I did it**
To be in sync with new logic and master code. gAsicInstance is needed for multi-asic platforms which gets missed as sai_redis library expects SAI_REDIS_SWITCH_ATTR_CONTEXT  as the last attribute,

**How I verified it**
Get a multi-asic platform up.