Fix issue: accumulative headroom can exceed limit in rare scenario (#2020)
- What I did
The egress mirror size is not taken into account when checking the accumulative headroom size, which causes the estimated accumulate headroom size to be greater than the real one and unable to be applied.

- Why I did it
Bug fix

- How I verified it
Run regression test.

- Details if related
An example to explain the logic .
The current accumulative headroom is 90k, egress mirror headroom is 10k, the ASIC limitation of the port is 100k.
Now a user wants to change the port configuration which will make the accumulative headroom to be increased by 10k.
Now the accumulative headroom estimated by buffer manager is 100k, which doesn’t exceed the limitation. However in the ASIC the real accumulative headroom is 110k which does.