[cbf] Fix max FC value (#2049)
What I did
Fixed the maximum FC value allowed by a switch.

Why I did it
There was a bit of confusion when I first wrote this, thinking the attribute should return the maximum FC value allowed by the switch, not the maximum number of FC classes. Because of this, the actual maximum FC value allowed is one less than the current one because we're counting from 0. As such, if the switch reports the maximum number of FCs is 255, the actual FC value must be in the 0-254 range.

How I verified it
Updated the existing UTs