[202012][pfc_detect] fix RedisReply errors (#2078)
- What I did
It could be that queue counters are populated to FLEX COUNTER DB before the COUNTER_QUEUE_INDEX_MAP is populated which is in accordance to PortsOrch implementation. Fixing it by replacing the order - first the COUNTER_QUEUE_INDEX_MAP then the FLEX COUNTER DB might produce other issues if readers start reading the map and try to find the corresponding entry in COUNTERS DB. Such an order is also in alignment with other maps and counters. So seems like fixing it in the reader is more suitable fix.

Errors observed only for short period of time at system start when PFC WD is enabled.

- Why I did it
If queue_index or port_id is not defined don't run the rest of the LUA script logic.

- How I verified it
Running it on the switch and verify no errors.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>