[chassis][syncd][sai] Adjusting response timeout during syncd init (#2159)
In VOQ based chassis where syncd uses VOQ SAI, if there are large
number of front panel ports, SAI takes more than 1 minutes to complete
the switch create initialization. Because of this, the switch create
request sent by orchagent is not getting response within the default
response wait time of 1 minute. So the orchagent declares switch create
failure and crashes.

To fix this, in orchagent, the syncd response time out is set to 5
minutes for line (voq) card and 10 minutes for supervisor (fabric) card
before sending request for switch create and is set back to default wait
time after the switch create.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>