[QosOrch] The notifications cannot be drained in QosOrch in case the first one needs to retry (#2206)
What I did
Bugfix: The notifications cannot be drained in QosOrch in case the first one needs to retry

Why I did it
Bugfix

How I verified it
Mock test and manual test.

Details if related
This is because both doTask and the table handlers take consumer as the argument.
In doTask, each notification is iterated in the main loop and passed to table handlers. However, consumer is passed to table handlers, which makes it unable to access the rest notifications following the first one. In case the first one needs to retry and remains in m_toSync, the rest notifications do not have opportunity to run.

Signed-off-by: Stephen Sun <stephens@nvidia.com>