[macsecorch]: MACsec with pfc (#2095)
What I did
Add PFC supported in the MACsec orchagent

Why I did it
Due to the different limitation to MACsec with PFC between different platform, we should have a configuration for trying to adapt with different neighbor devices.

How I verified it
Arista have verified this feature in the 7280 platform. They tried to sent the PFC packets under different configuration and to observe the change of PFC counter for verifying whether the configuration is available.