[201911][pfcwd] Avoid ingress drop by not attaching zero profiles when pfc storm is detected (#2279)
What I did
According to the current pfcwd detection logic on certain platforms, when wd fires, we create an ingress zero pool, ingress zero profile, egress zero pool, egress zero profile (if not already created) and then attach the ingress zero profile to the ingress pg and egress zero profile to the egress queue. As a result traffic ingressing that port/pg and egressing that port/queue will get dropped.

The current changes are done to avoid dropping traffic that is ingressing the port/pg that is in storm. The code changes in this PR avoid creating the ingress zero pool and profile and not attach any zero profile to the ingress pg when pfcwd is triggered

How I verified it
Modified the pfcwd func tests (Azure/sonic-mgmt#5665) and testcase passed on Mellanox platform

Signed-off-by: Neetha John <nejo@microsoft.com>