[tests] [asan] add graceful stop flag (#2347)
- What I did
Added a new flag to DVS tests

- Why I did it
Currently, when running the tests with ASAN-enabled image, leak reports are not generated. The reason is that dvs.destroy() (via 'ctn.remove(force=True)') uses SIGKILL to stop the container. To address this, a new flag is added.
When the new flag is set, the swss processes are gracefully stopped (via SIGTERM).
So ASAN reports can be generated as a result of DVS tests run

- How I verified it
Run the tests with --graceful-stop, observe that swss processes are stopped via SIGTERM

Signed-off-by: Yakiv Huryk <yhuryk@nvidia.com>