[routesync] Fix for stale dynamic neighbor (#2553)
* [routesync] Fix for stale dynamic neighbor

The changes are for fixing stale neighbor in the ASIC_DB and data path
when eBGP neighbors are shutdown and neighbors are flushed. The problem
is described in issue: https://github.com/sonic-net/sonic-buildimage/issues/12442
The root cause of this issue is due to not deleing the route from the
ASIC_DB when the route's next hop is on eth0 or docker0 interface. The solution is
to delete the route entry from ASIC_DB instead of just returning when the route's
next hop is on the interface eth0 or docker0


This commit fixes the warm restart unit test failure. When the route with
only nh on eth0 or docker0 is removed and if the route is the default
route, orchagent sends "add" black hole route to the syncd. So the ASIC
DB gets n hset message. When this happens during warm restart, the unit
test identifies this as unwanted setting and the unit test fails. To fix
this issues, the route delete is sent only if the warm restart is not in
progress. This is done following the same warm restart handling approach
used for route delete in other palces.

Signed-off-by: vedganes <veda.ganesan@nokia.com>